+++ tests/test_expression.py
@@ -352,7 +352,7 @@ def test_canonical():
     b2 = claripy.BoolS('b')
     c2 = claripy.BoolS('c')
 
-    assert x1.canonicalize()[-1] is x2.canonicalize()[-1]
+    assert x1.canonicalize() is x2.canonicalize()
 
     y1 = claripy.If(claripy.And(b1, c1), x1, ((x1+x1)*x1)+1)
     y2 = claripy.If(claripy.And(b2, c2), x2, ((x2+x2)*x2)+1)
@@ -362,7 +362,7 @@ def test_canonical():
 
     assert frozenset.union(*[a.variables for a in y1.recursive_leaf_asts]) == one_names
     assert frozenset.union(*[a.variables for a in y2.recursive_leaf_asts]) == two_names
-    assert y1.canonicalize()[-1] is y2.canonicalize()[-1]
+    assert y1.canonicalize() is y2.canonicalize()
 
 def test_depth():
     x1 = claripy.BVS('x', 32)
